requirements.txt
torch
torchvision
tqdm
pillow
opencv-python
keyboard
